feat(account): improve with TDD account create logic, generated with docker DB connection for testing, generated CI

+++ account-service/src/test/java/com/neowallet/accounts/domain/AccountTest.java
@@ -1,6 +1,9 @@
 package com.neowallet.accounts.domain;
 
 import org.junit.jupiter.api.Test;
+
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class AccountTest {
@@ -8,14 +11,14 @@ class AccountTest {
     @Test
     void shouldCreateAccountWithInitialBalance() {
         // Aquí fallará porque la clase Account aún no existe (Red phase de TDD)
-        Account account = new Account("123", 100.0);
-        assertEquals(100.0, account.getBalance());
+        Account account = new Account("123", BigDecimal.valueOf(100.0));
+        assertEquals(BigDecimal.valueOf(100.0), account.getBalance());
     }
 
     @Test
     void shouldThrowExceptionWhenBalanceIsNegative() {
         assertThrows(IllegalArgumentException.class, () -> {
-            new Account("123", -50.0);
+            new Account("123", BigDecimal.valueOf(-50.0));
         });
     }
 }
